fixed event card text formatting
